feat: display filter only when there are community credits
