+++ web-registry/src/pages/Projects/hooks/useProjectsSellOrders.types.ts
@@ -1,6 +1,15 @@
+import { ProjectWithOrderData } from '../Projects.types';
+
 export type GECKO_PRICES = {
   regenPrice?: number;
   eeurPrice?: number;
   usdcPrice?: number;
   evmosPrice?: number;
 };
+
+export interface ProjectsSellOrders {
+  projectsWithOrderData: ProjectWithOrderData[];
+  projectsCount?: number;
+  loading: boolean;
+  hasCommunityProjects: boolean;
+}
